refactor(export): extract renderItemWithEffects from renderFrame
Moves the ~168-line per-item render helper (transform + preview overrides +
effects across direct-GPU / preview-GPU / pooled-Canvas2D paths) into
frame-render-tasks.ts behind a FrameItemRenderDeps object. renderFrame keeps a
thin delegating wrapper preserving the 7-arg signature, so its 3 call sites are
unchanged. Body is verbatim; GPU-pipeline lazy-init still mutates the shared
itemRenderContext reference.

client-render-engine.ts: 2375 -> 2229 lines. Prunes 8 now-orphaned imports.
