+++ src/features/export/utils/frame-render-tasks.ts
@@ -1,9 +1,35 @@
+import type { TimelineItem } from '@/types/timeline'
+import type { ItemKeyframes } from '@/types/keyframe'
+import type { ItemEffect } from '@/types/effects'
+import type { ResolvedTransform } from '@/types/transform'
+import { createLogger } from '@/shared/logging/logger'
 import { doesMaskAffectTrack } from '@/shared/utils/mask-scope'
+import { getAnimatedTransform } from './canvas-keyframes'
+import { resolveAnimatedColorEffects } from '@/features/export/deps/keyframes'
+import {
+  combineEffects,
+  getAdjustmentLayerEffects,
+  renderEffectsFromMaskedSource,
+  type AdjustmentLayerWithTrackOrder,
+} from './canvas-effects'
+import { hasCornerPin } from '@/features/export/deps/composition-runtime'
 import type { CanvasPool } from './canvas-pool'
-import type { PreparedMask } from './canvas-masks'
+import type { MaskCanvasSettings, PreparedMask } from './canvas-masks'
 import type { ActiveTransition } from './canvas-transitions'
-import { renderTransitionToCanvas, type ItemRenderContext } from './canvas-item-renderer'
+import {
+  renderItem,
+  renderItemGpuEffectsToTexture,
+  renderPreviewVideoGpuEffectsToCanvas,
+  renderTransitionToCanvas,
+  type CanvasSettings,
+  type ItemRenderContext,
+} from './canvas-item-renderer'
+import type { GpuPipelineManager } from './gpu-pipeline-manager'
 import type { RenderedTaskResult } from './frame-mask-helpers'
+
+function getLog() {
+  return createLogger('ClientRenderEngine')
+}
 
 /**
  * Renders a transition to a pooled Canvas2D surface (the non-GPU fallback path),
@@ -33,4 +59,212 @@ export async function renderTransitionFallbackCanvas(
     transitionMasks,
   )
   return { source: trCanvas, poolCanvases: [trCanvas] }
+}
+
+export interface FrameItemRenderDeps {
+  frame: number
+  canvasSettings: CanvasSettings
+  maskSettings: MaskCanvasSettings
+  renderMode: 'export' | 'preview'
+  activeMasks: PreparedMask[]
+  adjustmentLayers: AdjustmentLayerWithTrackOrder[]
+  gpu: GpuPipelineManager
+  itemRenderContext: ItemRenderContext
+  canvasPool: CanvasPool
+  getCurrentItem: <TItem extends TimelineItem>(item: TItem) => TItem
+  getCurrentKeyframes: (itemId: string) => ItemKeyframes | undefined
+  getPreviewTransformOverride?: (itemId: string) => Partial<ResolvedTransform> | undefined
+  getPreviewCornerPinOverride?: (itemId: string) => TimelineItem['cornerPin'] | undefined
+  getPreviewEffectsOverride?: (itemId: string) => ItemEffect[] | undefined
+  getLiveItemSnapshot?: (itemId: string) => TimelineItem | undefined
+}
+
+/**
+ * Renders one timeline item with its transform, preview overrides, masks, and
+ * effects (direct-GPU, preview-GPU, or pooled Canvas2D paths). When `deferred`
+ * is true the rendered surface/texture is returned for later compositing;
+ * otherwise it is drawn straight to `targetCtx`. Extracted verbatim from
+ * `renderFrame`; mutates `deps.itemRenderContext.gpuPipeline` lazily.
+ */
+export async function renderItemWithEffects(
+  baseItem: TimelineItem,
+  trackOrder: number,
+  deferred: boolean,
+  targetCtx: CanvasRenderingContext2D | OffscreenCanvasRenderingContext2D,
+  deps: FrameItemRenderDeps,
+  bakeMasks = true,
+  preferGpuTextureOutput = false,
+  allowDirectGpu = true,
+): Promise<RenderedTaskResult | null> {
+  const {
+    frame,
+    canvasSettings,
+    maskSettings,
+    renderMode,
+    activeMasks,
+    adjustmentLayers,
+    gpu,
+    itemRenderContext,
+    canvasPool,
+    getCurrentItem,
+    getCurrentKeyframes,
+    getPreviewTransformOverride,
+    getPreviewCornerPinOverride,
+    getPreviewEffectsOverride,
+    getLiveItemSnapshot,
+  } = deps
+
+  const item = getCurrentItem(baseItem)
+  // Get animated transform
+  const itemKeyframes = getCurrentKeyframes(item.id)
+  let transform = getAnimatedTransform(item, itemKeyframes, frame, canvasSettings)
+  if (renderMode === 'preview') {
+    const previewOverride = getPreviewTransformOverride?.(item.id)
+    if (previewOverride) {
+      transform = {
+        ...transform,
+        ...previewOverride,
+        cornerRadius: previewOverride.cornerRadius ?? transform.cornerRadius,
+      }
+    }
+  }
+
+  // Apply corner pin preview override during interactive drag
+  let effectiveItem = item
+  if (renderMode === 'preview') {
+    const cornerPinOverride = getPreviewCornerPinOverride?.(item.id)
+    if (cornerPinOverride !== undefined) {
+      effectiveItem = { ...item, cornerPin: cornerPinOverride }
+    }
+  }
+
+  // Get effects (preview override → item effects + adjustment layer effects)
+  const baseItemEffects =
+    (renderMode === 'preview' ? getPreviewEffectsOverride?.(item.id) : undefined) ??
+    effectiveItem.effects
+  const itemEffects = resolveAnimatedColorEffects(
+    baseItemEffects,
+    getCurrentKeyframes(effectiveItem.id),
+    frame - effectiveItem.from,
+  )
+  const adjEffects = getAdjustmentLayerEffects(
+    trackOrder,
+    adjustmentLayers,
+    frame,
+    renderMode === 'preview' ? getPreviewEffectsOverride : undefined,
+    renderMode === 'preview' ? getLiveItemSnapshot : undefined,
+    getCurrentKeyframes,
+  )
+  const combinedEffects = combineEffects(itemEffects, adjEffects)
+  const applicableMasks = activeMasks.filter((mask) =>
+    doesMaskAffectTrack(mask.trackOrder, trackOrder),
+  )
+  const renderMasks = bakeMasks ? applicableMasks : []
+
+  // NOTE: The HTMLVideoElement importExternalTexture path stays disabled because
+  // textureSampleBaseClampToEdge produces subtly different edge pixel values
+  // compared to canvas 2D's drawImage (different YUV→RGB conversion at
+  // chroma subsampling boundaries). Spatial effects like halftone amplify
+  // this into a visible bright edge. The standard canvas 2D → GPU path
+  // below handles video correctly with negligible extra cost (~1-2ms).
+
+  const canRenderDirectGpuEffects =
+    allowDirectGpu &&
+    preferGpuTextureOutput &&
+    gpu.texturePool &&
+    itemRenderContext.gpuPipeline &&
+    itemRenderContext.gpuMediaPipeline &&
+    renderMasks.length === 0 &&
+    combinedEffects.length > 0 &&
+    combinedEffects.every((effect) => effect.enabled && effect.effect.type === 'gpu-effect') &&
+    (effectiveItem.type === 'video' || effectiveItem.type === 'image')
+  if (canRenderDirectGpuEffects && gpu.texturePool) {
+    const outputTexture = gpu.texturePool.acquire(canvasSettings.width, canvasSettings.height)
+    let renderedDirect = false
+    try {
+      renderedDirect = await renderItemGpuEffectsToTexture(
+        effectiveItem,
+        transform,
+        combinedEffects,
+        frame,
+        itemRenderContext,
+        outputTexture,
+        gpu.texturePool,
+      )
+      if (renderedDirect) {
+        return { gpuTexture: outputTexture, poolCanvases: [] }
+      }
+    } finally {
+      if (!renderedDirect) {
+        gpu.texturePool.release(outputTexture)
+      }
+    }
+  }
+
+  if (renderMasks.length === 0 && combinedEffects.length > 0) {
+    const directGpuCanvas = renderPreviewVideoGpuEffectsToCanvas(
+      effectiveItem,
+      transform,
+      combinedEffects,
+      frame,
+      itemRenderContext,
+    )
+    if (directGpuCanvas) {
+      if (deferred) {
+        return { source: directGpuCanvas, poolCanvases: [] }
+      }
+      targetCtx.drawImage(directGpuCanvas, 0, 0)
+      return null
+    }
+  }
+
+  // === PERFORMANCE: Use pooled canvas instead of creating new one ===
+  const { canvas: itemCanvas, ctx: itemCtx } = canvasPool.acquire()
+
+  // Render based on item type
+  await renderItem(
+    itemCtx,
+    effectiveItem,
+    transform,
+    frame,
+    itemRenderContext,
+    0,
+    undefined,
+    renderMasks,
+  )
+
+  // Apply effects (per-item — GPU effects applied here for both preview and export)
+  if (combinedEffects.length > 0) {
+    const hasGpu = combinedEffects.some((e) => e.enabled && e.effect.type === 'gpu-effect')
+    if (hasGpu && !itemRenderContext.gpuPipeline) {
+      itemRenderContext.gpuPipeline = await gpu.ensureEffects()
+      if (!itemRenderContext.gpuPipeline) {
+        getLog().warn('GPU pipeline init failed — GPU effects will be skipped')
+      }
+    }
+    const { source, poolCanvases } = await renderEffectsFromMaskedSource(
+      canvasPool,
+      itemCanvas,
+      combinedEffects,
+      hasCornerPin(effectiveItem.cornerPin) ? [] : renderMasks,
+      frame,
+      maskSettings,
+      itemRenderContext.gpuPipeline,
+    )
+    canvasPool.release(itemCanvas)
+
+    if (deferred) {
+      return { source, poolCanvases }
+    }
+    targetCtx.drawImage(source, 0, 0)
+    for (const effectCanvas of poolCanvases) canvasPool.release(effectCanvas)
+    return null
+  }
+
+  if (deferred) {
+    return { source: itemCanvas, poolCanvases: [itemCanvas] }
+  }
+  targetCtx.drawImage(itemCanvas, 0, 0)
+  canvasPool.release(itemCanvas)
+  return null
 }
